Revert "Merge remote-tracking branch 'origin/utf8-to-temporal-cast' into utf8-to-temporal-cast"
This reverts commit 215990bf85e355cb203caa246b643ff7d1c14c19, reversing
changes made to c5459f1226218c09478fb86ffa41453dd62c80a7.

+++ crates/polars-ops/src/chunked_array/interpolate.rs
@@ -242,34 +242,26 @@ mod test {
     fn test_interpolate() {
         let ca = UInt32Chunked::new("", &[Some(1), None, None, Some(4), Some(5)]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Linear);
-        let out = out.f64().unwrap();
+        let out = out.u32().unwrap();
         assert_eq!(
             Vec::from(out),
-            &[Some(1.0), Some(2.0), Some(3.0), Some(4.0), Some(5.0)]
+            &[Some(1), Some(2), Some(3), Some(4), Some(5)]
         );
 
         let ca = UInt32Chunked::new("", &[None, Some(1), None, None, Some(4), Some(5)]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Linear);
-        let out = out.f64().unwrap();
+        let out = out.u32().unwrap();
         assert_eq!(
             Vec::from(out),
-            &[None, Some(1.0), Some(2.0), Some(3.0), Some(4.0), Some(5.0)]
+            &[None, Some(1), Some(2), Some(3), Some(4), Some(5)]
         );
 
         let ca = UInt32Chunked::new("", &[None, Some(1), None, None, Some(4), Some(5), None]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Linear);
-        let out = out.f64().unwrap();
+        let out = out.u32().unwrap();
         assert_eq!(
             Vec::from(out),
-            &[
-                None,
-                Some(1.0),
-                Some(2.0),
-                Some(3.0),
-                Some(4.0),
-                Some(5.0),
-                None
-            ]
+            &[None, Some(1), Some(2), Some(3), Some(4), Some(5), None]
         );
         let ca = UInt32Chunked::new("", &[None, Some(1), None, None, Some(4), Some(5), None]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Nearest);
@@ -284,11 +276,8 @@ mod test {
     fn test_interpolate_decreasing_unsigned() {
         let ca = UInt32Chunked::new("", &[Some(4), None, None, Some(1)]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Linear);
-        let out = out.f64().unwrap();
-        assert_eq!(
-            Vec::from(out),
-            &[Some(4.0), Some(3.0), Some(2.0), Some(1.0)]
-        )
+        let out = out.u32().unwrap();
+        assert_eq!(Vec::from(out), &[Some(4), Some(3), Some(2), Some(1)])
     }
 
     #[test]
